requirements.txt
pyyaml
requests
psutil
python-dotenv
click==8.1.3
tabulate==0.8.10
click-spinner